airflow/requirements.txt
pmaw==2.1.3
apache-airflow-providers-google==7.0.0
pyarrow==8.0.0
pandas==1.4.2